chore: remove unused modules
The Libraries BOM has moved to the java-cloud-bom repository.
We have cloud-lts-bom still in this repository but it's not released
as part of the root project. Removing it from the root pom.xml.

The dashboard from the dashboard module is no longer used. Let's stop
the periodic build job.

The linkage-monitor directory is not released as part of the root
module.

This change also fixes the CI of the Gradle plugin.
The Gradle plugin is published to Gradle plugin portal, not Maven
Central. https://plugins.gradle.org/plugin/com.google.cloud.tools.linkagechecker